+++ tests/testthat/test-colwise-mutate.R
@@ -165,27 +165,6 @@ test_that("summarise_at removes grouping variables (#3613)", {
   expect_equal(names(res), c("g", "y"))
 })
 
-# Deprecated ---------------------------------------------------------
-
-test_that("_each() and _all() families agree", {
-  scoped_lifecycle_silence()
-  df <- data.frame(x = 1:3, y = 1:3)
-
-  expect_equal(summarise_each(df, list(mean)), summarise_all(df, mean))
-  expect_equal(summarise_each(df, list(mean), x), summarise_at(df, vars(x), mean))
-  expect_equal(summarise_each(df, list(mean = mean), x), summarise_at(df, vars(x), list(mean = mean)))
-  expect_equal(summarise_each(df, list(mean = mean), x:y), summarise_at(df, vars(x:y), list(mean = mean)))
-  expect_equal(summarise_each(df, list(mean), x:y), summarise_at(df, vars(x:y), mean))
-  expect_equal(summarise_each(df, list(mean), z = y), summarise_at(df, vars(z = y), mean))
-
-  expect_equal(mutate_each(df, list(mean)), mutate_all(df, mean))
-  expect_equal(mutate_each(df, list(mean), x), mutate_at(df, vars(x), mean))
-  expect_equal(mutate_each(df, list(mean = mean), x), mutate_at(df, vars(x), list(mean = mean)))
-  expect_equal(mutate_each(df, list(mean = mean), x:y), mutate_at(df, vars(x:y), list(mean = mean)))
-  expect_equal(mutate_each(df, list(mean), x:y), mutate_at(df, vars(x:y), mean))
-  expect_equal(mutate_each(df, list(mean), z = y), mutate_at(df, vars(z = y), mean))
-})
-
 test_that("group_by_(at,all) handle utf-8 names (#3829)", {
   with_non_utf8_encoding({
     name <- get_native_lang_string()
@@ -352,14 +331,8 @@ test_that("colwise mutate handle named chr vectors", {
 })
 
 test_that("colwise verbs soft deprecate quosures (#4330)", {
-  with_lifecycle_errors({
-    expect_error(
-      mutate_at(mtcars, vars(mpg), quo(mean(.)))
-    )
-    expect_error(
-      summarise_at(mtcars, vars(mpg), quo(mean(.)))
-    )
-  })
+  expect_warning(mutate_at(mtcars, vars(mpg), quo(mean(.))), "quosure")
+  expect_warning(summarise_at(mtcars, vars(mpg), quo(mean(.))), "quosure")
 
   expect_equal(
     transmute_at(mtcars, vars(mpg), ~. > mean(.)),