+++ tests/testthat/test-deprec-combine.R
@@ -1,8 +1,7 @@
-context("combine")
+setup(options(lifecycle_verbosity = "quiet"))
+teardown(options(lifecycle_verbosity = NULL))
 
 test_that("combine handles NULL (#1596, #3365)", {
-  scoped_lifecycle_silence()
-
   expect_equal(combine(list(NULL, 1, 2)), c(1, 2))
   expect_equal(combine(list(1, NULL, 2)), c(1, 2))
   expect_equal(combine(list(1, 2, NULL)), c(1, 2))
@@ -13,8 +12,6 @@ test_that("combine handles NULL (#1596, #3365)", {
 })
 
 test_that("combine complains about incompatibilites", {
-  scoped_lifecycle_silence()
-
   expect_error(
     combine("a", 1),
     class = "vctrs_error_incompatible_type"
@@ -26,14 +23,11 @@ test_that("combine complains about incompatibilites", {
 })
 
 test_that("combine works with input that used to fail (#1780)", {
-  scoped_lifecycle_silence()
   no <- list(alpha = letters[1:3], omega = letters[24:26])
   expect_equal(combine(no), unlist(no, use.names = FALSE))
 })
 
 test_that("combine works with NA and logical (#2203)", {
-  scoped_lifecycle_silence()
-
   # NA first
   expected_result <- c(NA, TRUE, FALSE, NA, TRUE)
   works1 <- combine(list(NA, TRUE, FALSE, NA, TRUE))
@@ -51,8 +45,6 @@ test_that("combine works with NA and logical (#2203)", {
 })
 
 test_that("combine works with NA and integers (#2203)", {
-  scoped_lifecycle_silence()
-
   works <- combine(list(1L, 2L, NA, 4L))
   expect_equal(works, c(1L, 2L, NA, 4L))
   works <- combine(list(1L, 2L, c(NA, NA), 4L))
@@ -60,8 +52,6 @@ test_that("combine works with NA and integers (#2203)", {
 })
 
 test_that("combine works with NA and factors (#2203)", {
-  scoped_lifecycle_silence()
-
   # NA first
   fac <- factor(c("a", "c", NA, "b"), levels = letters[1:3])
   expected_result <- fac[c(3, 1, 3, 2)]
@@ -87,8 +77,6 @@ test_that("combine works with NA and factors (#2203)", {
 })
 
 test_that("combine works with NA and double (#2203)", {
-  scoped_lifecycle_silence()
-
   # NA first
   works <- combine(list(NA, 1.5, 2.5, NA, 4.5))
   expect_equal(works, c(NA, 1.5, 2.5, NA, 4.5))
@@ -101,8 +89,6 @@ test_that("combine works with NA and double (#2203)", {
 })
 
 test_that("combine works with NA and characters (#2203)", {
-  scoped_lifecycle_silence()
-
   # NA first
   works <- combine(list(NA, "a", "b", "c", NA, "e"))
   expect_equal(works, c(NA, "a", "b", "c", NA, "e"))
@@ -116,8 +102,6 @@ test_that("combine works with NA and characters (#2203)", {
 
 
 test_that("combine works with NA and POSIXct (#2203)", {
-  scoped_lifecycle_silence()
-
   # NA first
   works <- combine(list(
     NA, as.POSIXct("2010-01-01"), as.POSIXct("2010-01-02"),
@@ -148,8 +132,6 @@ test_that("combine works with NA and POSIXct (#2203)", {
 })
 
 test_that("combine works with NA and Date (#2203)", {
-  scoped_lifecycle_silence()
-
   # NA first
   expected_result <- as.Date("2010-01-01") + c(NA, 1, 2, NA, 4)
   expect_equal(combine(as.list(expected_result)), expected_result)
@@ -183,8 +165,6 @@ test_that("combine works with NA and Date (#2203)", {
 
 
 test_that("combine works with NA and complex (#2203)", {
-  scoped_lifecycle_silence()
-
   # NA first
   expected_result <- c(NA, 1 + 2i)
   works1 <- combine(list(NA, 1 + 2i))
@@ -210,8 +190,6 @@ test_that("combine works with NA and complex (#2203)", {
 })
 
 test_that("combine works with integer64 (#1092)", {
-  scoped_lifecycle_silence()
-
   expect_equal(
     combine(bit64::as.integer64(2^34), bit64::as.integer64(2^35)),
     bit64::as.integer64(c(2^34, 2^35))
@@ -219,8 +197,6 @@ test_that("combine works with integer64 (#1092)", {
 })
 
 test_that("combine works with difftime", {
-  scoped_lifecycle_silence()
-
   expect_equal(
     combine(as.difftime(1, units = "mins"), as.difftime(1, units = "hours")),
     as.difftime(c(60, 3600), units = "secs")
@@ -252,8 +228,6 @@ test_that("combine works with difftime", {
 })
 
 test_that("combine uses tidy dots (#3407)", {
-  scoped_lifecycle_silence()
-
   chunks <- list(1,2,3)
   expect_equal(combine(!!!chunks), c(1,2,3))
 })